refactor:correct folder and variable names
