+++ task19/index.html
@@ -14,7 +14,7 @@
   <body class="p-16 bg-gray-300 font-sans tracking-widest">
     <div id="js-slideShowContainer" class="slideshow-container max-w-full mx-auto mb-12">
       <div id="js-slideShowItem" class="slideshow-item flex justify-between items-center">
-        <div id="js-slideShowWrapper" class="slideshow-wrapper w-8/12 h-96 mx-auto bg-white list-none relative overflow-hidden">
+        <div id="js-slideShowWrapper" class="slideshow-wrapper w-8/12 h-96 mx-auto bg-white relative overflow-hidden">
           <ul id="js-slideImageList" class="slide-image-list"></ul>
         </div>
       </div>
